New Edfited Proposal
The proposal was edited as the BB-AI was not functional and also the device was turning was hot and not connecting to the internet and since there was no time. Thus we opted to go with BB-WI and the proposal was required to be updated. For more information refer Issue #6 (https://github.com/zrajani/iLETTERBOX_CAPSTONE_PROJECT_EMBT_F2020/issues/6)